requirements.txt
json
pandas
plotly
locale
streamlit
scikit-learn
python-dotenv
requests